app/middleware/error_handler.py
"""Global error handler middleware — ensures consistent error response format."""
import logging
import traceback
from fastapi import Request
from fastapi.responses import JSONResponse
from starlette.middleware.base import BaseHTTPMiddleware

log = logging.getLogger(__name__)


class ErrorHandlerMiddleware(BaseHTTPMiddleware):
    """Catch unhandled exceptions and return structured error responses."""

    async def dispatch(self, request: Request, call_next):
        try:
            return await call_next(request)
        except Exception as exc:
            request_id = getattr(request.state, "request_id", "unknown")
            log.exception(
                "Unhandled error: request_id=%s path=%s error=%s",
                request_id, request.url.path, str(exc),
            )
            return JSONResponse(
                status_code=500,
                content={
                    "error": "internal_server_error",
                    "message": str(exc),
                    "request_id": request_id,
                    "path": request.url.path,
                },
            )
