app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
numpy>=1.26.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
databricks-sdk>=0.56.0
requests>=2.31.0
scipy>=1.11.0
fpdf2>=2.7.0
